requirements.txt
py-cord==2.5.0
aiohttp==3.9.5
apscheduler==3.10.4
beautifulsoup4==4.12.3
dateutils==0.6.12
python-dotenv==1.0.1
requests==2.32.3
yt-dlp==2024.5.27
pynacl==1.5.0